+++ crates/authly-common/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- added `GetConfiguration` to authly service proto, replaces `GetResourcePropertyMappings`.
+
 ### Added
 - public method to get peer service entity from MTLSMiddleware when not used with HTTP